// appfront/src/app/api/models/conversation-summary-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ParticipantSummaryResponse } from '../models/participant-summary-response';
export interface ConversationSummaryResponse {
  conversationId?: string;
  courseId?: string;
  courseTitle?: string;
  lastMessageContent?: string;
  lastMessageSentAt?: string;
  recipient?: ParticipantSummaryResponse;
  type?: 'DIRECT' | 'COURSE_SUPPORT';
  unreadCount?: number;
  updatedAt?: string;
}