// appfront/src/app/api/models/message-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface MessageResponse {
  attachmentUrl?: string;
  content?: string;
  idConversation?: string;
  idMessage?: string;
  isMine?: boolean;
  senderId?: string;
  senderName?: string;
  sentAt?: string;
  status?: 'SENT' | 'DELIVERED' | 'READ';
}